requirements.txt
flask
deepface
psycopg2-binary
redis
tf-keras
hdbscan